requirements.txt
google-genai>=0.3.0
pydantic>=2.0.0
google-cloud-storage>=2.10.0
google-cloud-secret-manager>=2.16.0
flask>=3.0.0
